fixed some unsigned/signed mismatch issues

+++ src/components/denoising/tv-denoising/TVDenoisingComponent.cpp
@@ -40,7 +40,7 @@ namespace components {
         void TVDenoisingComponent::applyDenoising() {
             float loss_smoothed{ 0.0f };
 
-            uint64_t counter { 1U };
+            uint32_t counter { 1U };
             while (true) {
                 float loss{ evalLossAndGrad() };
 
@@ -49,7 +49,7 @@ namespace components {
                 }
 
                 loss_smoothed = loss_smoothed * loss_smoothing_beta + loss * (1.0f - loss_smoothing_beta);
-                float loss_smoothed_debiased{ loss_smoothed / (1.0f - static_cast<float>(std::pow(loss_smoothing_beta, counter))) };
+                float loss_smoothed_debiased{ loss_smoothed / (1.0f - static_cast<float>(std::pow(loss_smoothing_beta, static_cast<float>(counter)))) };
                 if (counter > 1 && loss_smoothed_debiased / loss < 1.0f + tolerance) {
                     if (ENABLE_LOGGING) {
                         std::cout << "Converged after " << counter << " iterations with loss: " << loss_smoothed_debiased << std::endl;
